backend/requirements.txt
# FastAPI and server
fastapi>=0.68.0
uvicorn>=0.15.0

# Database
sqlalchemy>=1.4.23
psycopg2-binary>=2.9.1
asyncpg>=0.27.0
alembic>=1.7.1  # for database migrations

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5

# Environment and settings
python-dotenv>=0.19.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# AI/ML
openai>=0.27.0

# Caching
redis>=5.0.0

# Testing
pytest>=6.2.5
pytest-asyncio>=0.15.1
httpx>=0.18.2  # for async HTTP testing

# Logging
loguru>=0.6.0
python-json-logger>=2.0.7 